have now finished problem 3.
Problem 3 solved in similar fashion to problem 2, using nested data and then applying a written function to solve for sample mean and perform a t-test to determine differences between sampled and true mean.

This data was tidied and mutated to produce relevaant information that can displayed graphically to indicate that there is more freuqent significant difference between the true and sample mean as the true mean rises.